Initial commit of the knead-wikidough test suite
This is the first commit of the knead-wikidough test suite for the
production testing of Wikidough. Currently it has sixteen tests to
validate the TLS and DNS configuration of Wikidough and the interaction
between its dnsdist and pdns-recursor components.

This is a work-in-progress and more tests are required to cover all
Wikidough configurations.

Bug: T267424
Change-Id: Iac633479176336a248e69fc5d7d18637193d7c64